created full framework for first naive circuit
Project consists of following parts:
- Dockerfile + compose.yml for Circom and SnarkJS
    - docker-compose provides separate command for compilation of
      circuit without needing to enter the container itself
- Node App
    - provides user interface to create/verify/test proofs
    - basic functionality provided by SnarkJS library
    - also measures execution time --> needed for benchmarking
- Pattern Generator (Py)
    - python script for generating test patterns
        - can then extract many test cases from one test pattern
        - output as JSON --> can be parsed by Node app
- Benchmarking (Py)
    - python script that uses all of the above components to run
      automatic benchmarking with various sizes and compilation flags
    - prints results as json

The also is a lot of documentation already for the node app, the pattern
generator and the benchmarking script. Python code is documented using
sphinx, JS is documented using jsdoc. But there's supposed to be a way
to also document JS with sphinx.

Most components have extensive error handling, nice commandline
interface, usage messages, etc.

Also includes benchmarking of first circuit:
- For input sizes between 16 pixels and 256
- Takes much too long
- Circuit is way too large
- For larger sizes, circom process runs out of stack

What's missing is a combined overview of the whole project, but i will
probably put this into the report.